requirements.txt
feedparser
hydralit_components
matplotlib
nltk
numpy
pandas
Pillow
plotly
scikit_learn
streamlit
streamlit_option_menu
tensorflow
textblob
tweepy
vaderSentiment
yfinance
